requirements.txt
pandas
requests
pyarrow
psycopg2-binary